+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Moved `CountingWriter` from lzma2 to sevenz crate, since it was an internal implementation detail of the sevenz crate.
+
 ## 0.12.1 - 2025-03-10
 
 ### Fixed